Resolve Safari bug
Originally, when you clicked outside the Toast, the Toast would
automatically close. However, on Safari, it registered the action of the
click that "triggers" the toast as the same action to close the Toast,
which prevented the Toast from being visible to the user. (it would
open/close so quickly it wasn't noticeable to the eye)